Update payment intents.
