+++ Sources/StripeKit/Core Resources/PaymentIntents/PaymentIntent.swift
@@ -52,23 +52,27 @@ public struct StripePaymentIntent: StripeModel {
     /// Set of key-value pairs that you can attach to an object. This can be useful for storing additional information about the object in a structured format.
     public var metadata: [String: String]?
     /// If present, this property tells you what actions you need to take in order for your customer to fulfill a payment using the provided source.
-    public var StripePaymentIntentNextAction: String?
+    public var nextAction: StripePaymentIntentNextAction?
     /// The account (if any) for which the funds of the PaymentIntent are intended. See the PaymentIntents Connect usage guide for details.
     public var onBehalfOn: String?
     /// ID of the payment method used in this PaymentIntent.
     public var paymentMethod: String?
+    /// Payment-method-specific configuration for this PaymentIntent.
+    public var paymentMethodOptions: StripePaymentIntentPaymentMethodOptions?
     /// The list of payment method types (e.g. card) that this PaymentIntent is allowed to use.
     public var paymentMethodTypes: [String]?
     /// Email address that the receipt for the resulting payment will be sent to.
     public var receiptEmail: String?
     /// ID of the review associated with this PaymentIntent, if any.
     public var review: String?
+    /// Indicates that you intend to make future payments with this PaymentIntent’s payment method. If present, the payment method used with this PaymentIntent can be attached to a Customer, even after the transaction completes. Use on_session if you intend to only reuse the payment method when your customer is present in your checkout flow. Use off_session if your customer may or may not be in your checkout flow. For more, learn to save card details after a payment. Stripe uses setup_future_usage to dynamically optimize your payment flow and comply with regional legislation and network rules. For example, if your customer is impacted by SCA, using off_session will ensure that they are authenticated while processing this PaymentIntent. You will then be able to collect off-session payments for this customer.
+    public var setupFutureUsage: String?
     /// Shipping information for this PaymentIntent.
     public var shipping: StripeShippingLabel?
-    /// ID of the source used in this PaymentIntent.
-    public var source: String?
-    /// Extra information about a PaymentIntent. This will appear on your customer’s statement when this PaymentIntent succeeds in creating a charge.
+    /// For non-card charges, you can use this value as the complete description that appears on your customers’ statements. Must contain at least one letter, maximum 22 characters.
     public var statementDescriptor: String?
+    /// Provides information about a card payment that customers see on their statements. Concatenated with the prefix (shortened descriptor) or statement descriptor that’s set on the account to form the complete statement descriptor. Maximum 22 characters for the concatenated descriptor.
+    public var statementDescriptorSuffix: String?
     /// Status of this PaymentIntent, one of `requires_payment_method`, `requires_confirmation`, `requires_action`, `processing`, `requires_capture`, `canceled`, or `succeeded`.
     public var status: StripePaymentIntentStatus?
     /// The data with which to automatically create a Transfer when the payment is finalized. See the PaymentIntents Connect usage guide for details.
@@ -133,4 +137,26 @@ public struct StripePaymentIntentsList: StripeModel {
     public var hasMore: Bool?
     public var url: String?
     public var data: [StripePaymentIntent]?
+}
+
+public struct StripePaymentIntentPaymentMethodOptions: StripeModel {
+    /// If the PaymentIntent’s `payment_method_types` includes `card`, this hash contains the configurations that will be applied to each payment attempt of that type.
+    public var card: StripePaymentIntentPaymentMethodOptionsCard?
+}
+
+public struct StripePaymentIntentPaymentMethodOptionsCard: StripeModel {
+    /// Installment details for this payment (Mexico only). For more information, see the installments integration guide.
+    public var installments: StripePaymentIntentPaymentMethodOptionsCardInstallments?
+    /// We strongly recommend that you rely on our SCA Engine to automatically prompt your customers for authentication based on risk level and other requirements. However, if you wish to request 3D Secure based on logic from your own fraud engine, provide this option. Permitted values include: `automatic` or `any`. If not provided, defaults to `automatic`. Read our guide on manually requesting 3D Secure for more information on how this configuration interacts with Radar and our SCA Engine.
+    public var requestThreeDSecure: String?
+}
+
+public struct StripePaymentIntentPaymentMethodOptionsCardInstallments: StripeModel {
+    // TODO: - Rename the charge installment plan.
+    /// Installment plans that may be selected for this PaymentIntent.
+    public var availablePlans: [StripeChargePaymentDetailsCardInstallmentPlan]?
+    /// Whether Installments are enabled for this PaymentIntent.
+    public var enabled: Bool?
+    /// Installment plan selected for this PaymentIntent.
+    public var plan: StripeChargePaymentDetailsCardInstallmentPlan?
 }
